+++ tests/e2e/specs/front-end/test-search-browse-sort.js
@@ -1,0 +1,6 @@
+/*
+ * Internal dependencies
+ */
+const { runSearchBrowseSortTest } = require( '@woocommerce/e2e-core-tests' );
+
+runSearchBrowseSortTest();