+++ tests/e2e/core-tests/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 
 - Shopper Cart Apply Coupon
+- Shopper Search, Browse & Sort Products
 
 ## Fixed